+++ assets/vue/components/compression.vue
@@ -18,6 +18,27 @@
                                color="#008ec2"></toggle-button>
             </div>
         </div>
+
+        <!--CSS & JS Assets load toggle -->
+        <div class="field  is-fullwidth columns" >
+            <label class="label column has-text-grey-dark">
+                {{strings.toggle_cdn}}
+
+                <p class="is-italic has-text-weight-normal">
+                    {{strings.cdn_desc}}
+                </p>
+            </label>
+
+            <div class="column is-3 ">
+                <toggle-button :class="'has-text-dark'"
+                               v-model="cdnStatus"
+                               :disabled="this.$store.state.loading"
+                               :labels="{checked: strings.enabled, unchecked: strings.disabled}"
+                               :width="80"
+                               :height="25"
+                               color="#008ec2"></toggle-button>
+            </div>
+        </div>
         <!-- GIF replacement toggle button -->
         <div class="field  columns">
             <label class="label column has-text-grey-dark">
@@ -259,6 +280,15 @@
 
 		        }
 	        },
+			cdnStatus: {
+				set: function (value) {
+					this.showSave = true;
+					this.new_data.cdn = value ? 'enabled' : 'disabled';
+				},
+				get: function () {
+					return !(this.site_settings.cdn === 'disabled');
+				}
+			},
 			reportScriptStatus: {
 				set: function (value) {
 					this.showSave = true;