+++ assets/vue/components/general.vue
@@ -19,7 +19,9 @@
             </div>
 
         </div>
-        <!--Scale and lazy load toggle -->
+
+        <hr/>
+        <!--Lazy load toggle -->
         <div class="field  is-fullwidth columns" :class="{'is-field-disabled':isReplacerOff }">
             <label class="label column has-text-grey-dark">
                 {{strings.toggle_lazyload}}
@@ -39,28 +41,11 @@
                                color="#008ec2"></toggle-button>
             </div>
         </div>
-        <!--CSS & JS Assets load toggle -->
-        <div class="field  is-fullwidth columns" :class="{'is-field-disabled':isReplacerOff }">
-            <label class="label column has-text-grey-dark">
-                {{strings.toggle_cdn}}
 
-                <p class="is-italic has-text-weight-normal">
-                    {{strings.cdn_desc}}
-                </p>
-            </label>
+        <hr/>
 
-            <div class="column is-3 ">
-                <toggle-button :class="'has-text-dark'"
-                               v-model="cdnStatus"
-                               :disabled="this.$store.state.loading"
-                               :labels="{checked: strings.enabled, unchecked: strings.disabled}"
-                               :width="80"
-                               :height="25"
-                               color="#008ec2"></toggle-button>
-            </div>
-        </div>
         <!-- Clear Cache button -->
-        <div class="field  is-fullwidth columns ">
+        <div class="field  is-fullwidth columns " :class="{'is-field-disabled':isReplacerOff }">
             <label class="label column has-text-grey-dark">
                 {{strings.cache_title}}
 
@@ -99,6 +84,7 @@
                 showNotification: false,
                 showSave: false,
                 isReplacerOff: false,
+                isLazyLoadOff : false,
                 new_data: {},
 
             }
@@ -137,22 +123,13 @@
             lazyLoadStatus: {
                 set: function (value) {
                     this.showSave = true;
+                    this.isLazyLoadOff = !value;
                     this.new_data.lazyload = value ? 'enabled' : 'disabled';
                 },
                 get: function () {
                     return !(this.site_settings.lazyload === 'disabled');
-                }
-            },
-	        cdnStatus: {
-		        set: function (value) {
-			        this.showSave = true;
-			        this.new_data.cdn = value ? 'enabled' : 'disabled';
-		        },
-		        get: function () {
-			        return !(this.site_settings.cdn === 'disabled');
-		        }
+                } 
             }
-
         }
     }
 </script>